-- Work in progress --
Added new DataExchangeViewRep for changing the set id of the views.
Added the corresponding command to create the view rep.
For now changing the data SET is only possible for views (and not GLCanvasUser).
This needs to be added soon.
The newly assigned SET must fit to the view rep. An assignment of a wrong SET is not checked at the moment.
The data changing is tested for the 2D pathway view - because this is the only view at the moment that acts upon a SET.


git-svn-id: svn://localhost:8080/caleydo_main/trunk@516 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3