Added labels to the data exchanger ViewRep. The first drop down list is to select the view. In the second drop down list a new SET can be assigned to the selected view.
Started to solve BUGS.
Added comments in BUG_REPORT.txt and marked the solved issues as [SOLVED].

Added command (CmdViewLoadURLInHTMLBrowser) for loading URLs in browser (this was open issue in BUG_REPORT).
Registered command in CommandFactory.
BUT: The command cannot be used at the moment because the interface of the CommandManager needs to be changed a bit.

Updated JabRef file (two new papers are added).

git-svn-id: svn://localhost:8080/caleydo_main/trunk@520 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3