Selection update for pathways is working in BOTH DIRECTIONS at the same time (without neighborhood). (Minor problem: Previously highlighted vertices in 2D pathways are not set back on new selection).
Changed ISet to SetSelection in all selection relevant interfaces (EventPublisher, Mediator, etc.)
Because there is no reason to hand over a abstract ISet. There is no need to cast at the receiver.

FixedBUG with double insertion of EventMediators in a HashMap. That was the reason why the update was called twice.

git-svn-id: svn://localhost:8080/caleydo_main/trunk@561 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3