Updated and moved XML configuration files.
The CHI and CHI 4 Views setups are moved to folder "michael". Both are running inclusive heatmap!

All "dead" configuration files are moved to folder "old".

Unit testing XML configuration is moved to folder "shared/testing".

Note: After this update the Run configurations in local Eclipse installations must be altered to point to the new folders. This is the only thing that needs to be updated locally.

git-svn-id: svn://localhost:8080/caleydo_main/trunk@1088 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3