-- WORK IN PROGRESS -- Re-factoring of the JOGL canvas handling and management. It now uses the JOGl internal scheduling. This solution is much cleaner and even more powerful. Due to this fact a lot of code become useless and therefore removed.
A lot of cleanup is currently in progress (not fully completed yet).
Some classes (e.g. scatterplot 3D, etc) are currently commented out and need a update when they come back to life again.
Other things need to be done in the next days to complete the port. For example the viewing parameters must be stored in the jogl canvas forwarder.

git-svn-id: svn://localhost:8080/caleydo_main/trunk@1123 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3