Added context menu for load embedded pathways (ticket#311).
GUI bridge provides access to current display. The display can be used for async exec to the GUI thread.
This is needed to create e.g. message boxes in the core.

Replaced NA with NaN in Asslaber data file. The loading runs therefore without problems.
Still there are problems with the clinical use case which needs to be fixed (soon).

git-svn-id: svn://localhost:8080/caleydo_main/trunk@2049 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3