visdaemon: sending window access information from vis renderer to vis daemon. Visdaemon now only updates the user's links if it is in the list of target windows AND is accessible.
This is necessary, as it now also searches for selections in non-accessible windows. However, links should not be redrawn if the window content in non-accessible windows changes (e.g. the owner scrolls his window on the private display), as only arrows are rendered which are not affected by window scrolling. 

git-svn-id: svn://localhost:8080/caleydo_main/trunk@3215 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3